smartmirror_pi5/requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
pytz==2025.2
opencv-python==4.10.0.84
numpy==2.0.2
SpeechRecognition==3.10.4
PyAudio==0.2.14